chore: Add domain directory and move classes
