+++ src/main/kotlin/minesweeper/domain/Cell.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.domain
 
 interface Cell {
     val state: String
