+++ src/main/kotlin/minesweeper/domain/CleanCell.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.domain
 
 class CleanCell(override val state: String = "C") : Cell {
 }
