+++ src/main/kotlin/minesweeper/domain/MineCell.kt
@@ -1,3 +1,3 @@
-package minesweeper
+package minesweeper.domain
 
 class MineCell(override val state: String = "*") : Cell
